+++ Plugins/Core/Source/NexusCore/Private/NRandom.cpp
@@ -6,5 +6,5 @@
 #include "Math/NSeedGenerator.h"
 
 // We'll do a static initialization of the core random for now, but this eventually would be set during any sort of multiplayer or loading scenario.
-FNMersenneTwister FNRandom::Deterministic = FNMersenneTwister(FNSeedGenerator::SeedFromText(TEXT("NEXUS")));
+FNMersenneTwister FNRandom::Deterministic = FNMersenneTwister(FNSeedGenerator::SeedFromString(TEXT("NEXUS")));
 FRandomStream FNRandom::NonDeterministic = FRandomStream(FDateTime::Now().GetMillisecond());