+++ Plugins/Core/Source/NexusCore/Public/Math/NSeedGenerator.h
@@ -8,24 +8,32 @@ class NEXUSCORE_API FNSeedGenerator
 public:
 
 	/**
-	 * Determine if the provided seed is <i>valid</i>.
+	 * Determine if the provided seed is valid.
 	 * @param InHexSeed - The FString to check if it is valid
 	 * @return an indicator of the seeds' validity
 	 */
 	static bool IsValidHexSeed(const FString& InHexSeed);
 
 	/**
-	 * Returns a <i>valid</i> pseudo random seed to be used with FNMersenneTwister.
+	 * Returns a valid pseudo random seed to be used with FNMersenneTwister.
 	 * @return a numeric seed
 	 */
 	static uint64 RandomSeed();
+
+	/**
+	 * Returns a friendly seed with a specified number of blocks and block length.
+	 * @param BlockCount - The number of blocks to generate
+	 * @param BlockLength - The length of each block
+	 * @return A friendly seed string
+	 */
+	static FString RandomFriendlySeed(uint8 BlockCount = 3, uint8 BlockLength = 5);
 
 	/**
 	 * Returns a <i>valid</i> numeric seed to be used with the FNMersenneTwister.
 	 * @param InSeed - The desired FString to parse
 	 * @return a numeric seed
 	 */
-	static uint64 SeedFromText(const FString& InSeed);
+	static uint64 SeedFromString(const FString& InSeed);
 
 	/**
 	 * Returns a <i>valid</i> numeric seed to be used with the FNMersenneTwister.
@@ -51,9 +59,17 @@ private:
 	static int32 HexToInteger(const TCHAR Char);
 
 	/**
-	 * Removes all <i>non-valid</i> characters from the provided seed.
+	 * Removes all non-valid characters from the provided seed.
 	 * @param InHexSeed - The FString input seed
 	 * @return a cleaned-up version of the provided seed
 	 */
 	static FString SanitizeHexSeed(const FString& InHexSeed);
+
+	/**
+	 * Generate a random block of lowercase alpha characters to use with friendly seeds.
+	 * @param Length The number of characters to generate.
+	 * @return The readable block of characters.
+	 */
+	FORCEINLINE static FString RandomFriendlySeedBlock(uint8 Length);
+	
 };