+++ Plugins/ProcGen/Source/NexusProcGen/Public/NProcGenRegistry.h
@@ -50,7 +50,7 @@ private:
 	static void NotifyOfStateChange(UNProcGenOperation* Operation, ENProcGenOperationState NewState)
 	{
 		// We don't have a display name at registration, so the log for registration is done manually
-		UE_CLOG(NewState !=ENProcGenOperationState::PGOS_Registered, LogNexusProcGen, Log, TEXT("[%s] Changed State(%s)"), *Operation->GetDisplayName().ToString(), *UNProcGenOperation::GetStringFromState(NewState));
+		UE_CLOG(NewState != ENProcGenOperationState::PGOS_Registered, LogNexusProcGen, Log, TEXT("[%s] Changed State(%s)"), *Operation->GetDisplayName().ToString(), *UNProcGenOperation::GetStringFromState(NewState));
 		OnOperationStateChanged.Broadcast(Operation, NewState);
 	}